Added fees

+++ server/wolt22/src/main/kotlin/com/junction/wolt22/beans/DeliveryResponse/Coordinates.kt
@@ -1,8 +1,8 @@
 package com.junction.wolt22.beans.DeliveryResponse
 
-import kotlin.Int
+import kotlin.Double
 
 data class Coordinates(
-    val lat: Int,
-    val lon: Int
+    val lat: Double,
+    val lon: Double
 )
